v0 Update 94 -> Updated creating controls on AddControl
