+++ src/ISUF.UI/ViewModel/ModuleAddVMBase.cs
@@ -164,122 +164,22 @@ namespace ISUF.UI.ViewModel
             AddEditItem = currentItem;
         }
 
-        protected virtual IList<FrameworkElement> GetControlsFromForm()
+        protected virtual void SaveItem()
         {
-            return Classes.Functions.GetControlsByName(form, Classes.Constants.DATA_CONTROL_IDENTIFIER, true);
+            FillValuesFromFormIntoItem();
+            uiModule.AddItem(AddEditItem);
         }
 
         public void FillValuesFromFormIntoItem()
         {
-            var formControls = GetControlsFromForm();
-
-            FillValues(formControls);
-        }
-
-        protected virtual void SaveItem()
-        {
-            FillValuesFromFormIntoItem();
-
-            uiModule.AddItem(AddEditItem);
+            var formControls = FormDataMiner.GetControlsFromForm(form);
+            AddEditItem = FormDataMiner.FillValuesIntoProperty(formControls, AddEditItem);
         }
 
         private void SaveCloseItem()
         {
             SaveItem();
             CloseAddPane();
-        }
-
-        private void FillValues(IList<FrameworkElement> formControls)
-        {
-            var itemProps = typeof(T).GetProperties();
-
-            foreach (var formControl in formControls)
-            {
-                var formControlName = formControl.Name.Substring(0, formControl.Name.Length - Classes.Constants.DATA_CONTROL_IDENTIFIER.Length);
-                var itemProp = itemProps.FirstOrDefault(x => x.Name == formControlName);
-
-                if (itemProp == null)
-                    throw new Base.Exceptions.ArgumentException("None of controls not match property name.");
-
-                object value = GetValueFromControl(formControl);
-
-                if (value.GetType() != itemProp.PropertyType)
-                    value = ConvertValueToPropValue(value, itemProp);
-
-                itemProp.SetValue(AddEditItem, value);
-            }
-        }
-
-        private object ConvertValueToPropValue(object value, PropertyInfo prop)
-        {
-            object convertedValue;
-            bool conversionResult = true;
-
-            if (value.GetType() == typeof(string))
-                if (prop.PropertyType == typeof(int))
-                {
-                    conversionResult = int.TryParse(value as string, out int convertedNumber);
-                    convertedValue = convertedNumber;
-                }
-                else if (prop.PropertyType == typeof(byte))
-                {
-                    conversionResult = byte.TryParse(value as string, out byte convertedNumber);
-                    convertedValue = convertedNumber;
-                }
-                else if (prop.PropertyType == typeof(double))
-                {
-                    conversionResult = double.TryParse(value as string, out double convertedNumber);
-                    convertedValue = convertedNumber;
-                }
-                else if (prop.PropertyType == typeof(decimal))
-                {
-                    conversionResult = decimal.TryParse(value as string, out decimal convertedNumber);
-                    convertedValue = convertedNumber;
-                }
-                else if (prop.PropertyType == typeof(float))
-                {
-                    conversionResult = float.TryParse(value as string, out float convertedNumber);
-                    convertedValue = convertedNumber;
-                }
-                else if (prop.PropertyType == typeof(char))
-                    return value;
-                else
-                    throw new Base.Exceptions.NotSupportedException("Not supported type conversion between control and property. \n" +
-                    $"Value type: {value.GetType()}\n\n" +
-                    $"Target type: {prop.PropertyType}");
-            else if (value.GetType() == typeof(TimeSpan))
-                convertedValue = DateTime.Today + (TimeSpan)value;
-            else if (value.GetType() == typeof(DateTimeOffset))
-                convertedValue = ((DateTimeOffset)value).DateTime;
-            else
-                throw new Base.Exceptions.NotSupportedException("Not supported type conversion between control and property. \n" +
-                    $"Value type: {value.GetType()}\n" +
-                    $"Target type: {prop.PropertyType}");
-
-            if (!conversionResult)
-                throw new Base.Exceptions.UnsuccessfulConversionException("Conversion was unsuccessful. \n" +
-                    $"Value type: {value.GetType().ToString()}\n" +
-                    $"Target type: {prop.PropertyType.Name}");
-
-            return convertedValue;
-        }
-
-        private object GetValueFromControl(FrameworkElement formControl)
-        {
-            if (formControl.GetType() == typeof(TextBox))
-                return (formControl as TextBox).Text;
-            else if (formControl.GetType() == typeof(LinkedTablePresenterControl))
-                return (formControl as LinkedTablePresenterControl).GetSelectedIds();
-            else if (formControl.GetType() == typeof(LinkedTableSelectorControl))
-                return (formControl as LinkedTableSelectorControl).GetSelectedId();
-            else if (formControl.GetType() == typeof(CalendarDatePicker))
-                return (formControl as CalendarDatePicker).Date;
-            else if (formControl.GetType() == typeof(TimePicker))
-                return (formControl as TimePicker).Time;
-            else if (formControl.GetType() == typeof(CheckBox))
-                return (formControl as CheckBox).IsChecked;
-            else
-                throw new Base.Exceptions.NotSupportedException("Not supported type of control for getting value.");
         }
 
         /// <summary>
